+++ src/Fields/AttachmentOption.php
@@ -1,5 +1,5 @@
 <?php
-namespace Zawntech\WPAdminOptions;
+namespace AllegedWizard\WPAdminOptions\Fields;
 
 class AttachmentOption extends AbstractAdminOption
 {