+++ src/Fields/DurationOption.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Zawntech\WPAdminOptions;
+namespace AllegedWizard\WPAdminOptions\Fields;
 
 class DurationOption extends AbstractAdminOption
 {