+++ src/Fields/InputOption.php
@@ -1,4 +1,4 @@
 <?php
-namespace Zawntech\WPAdminOptions;
+namespace AllegedWizard\WPAdminOptions\Fields;
 
 class InputOption extends AbstractAdminOption {}